Implemented a hacky in-memory HardcoreRouter
We have the RoutingService which handles the grunt work of fetching data from
storage (in this case some in-memory maps) - and we should then be able to
replace that with a DB version.

We base the HardcoreRouter off of the ProtoRouter, with a few adjustments.
First we make `urlFor` async - this allows for async lookup from the DB in
future. Second, we update the handling of the `getDataForUrl` return value, so
that we can render templates. Both of these changes can be pulled into ProtoRouter.

This will be useful because 1. future prototypes will be easier and 2. the
differences between the prototypes will be smaller, and easier to comprehend.

We have a weird populateRoutingService method - which in reality would be
replaced by a database migration (to populate routes for resources) and maybe
some kind of reading from a routes.yaml for the homepage stuff.