Added support for custom URL assignment in Posts API
refs https://github.com/TryGhost/Arch/issues/64

We need a way to modify existing URLs in the HardcoreRouter's URL Service. Allowed to pass in a `url` property along with post's payload that overrides existing URL assigned to the post resource.

If the new URL clashes with any existing ones it gets re-assigned to a newly generated one.